% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff.R
\name{aff_read_key}
\alias{aff_read_key}
\title{FUNCTION_TITLE}
\usage{
aff_read_key(filename, key, key_length = 0)
}
\arguments{
\item{filename}{PARAM_DESCRIPTION}

\item{key}{PARAM_DESCRIPTION}

\item{key_length}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
