% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff.R
\name{aff_read_key_list}
\alias{aff_read_key_list}
\title{FUNCTION_TITLE}
\usage{
aff_read_key_list(filename, key_list, key_length = 0)
}
\arguments{
\item{filename}{PARAM_DESCRIPTION}

\item{key_list}{PARAM_DESCRIPTION}

\item{key_length}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 aff_read_key_list(filename = "piN_piN_diagrams.1700.tsrc04.keys",
 }
}
}
